Make request modal more generic
Make all modal inputs hidden and readonly by default, empty the header and footer and combine the previous two forms into one that contains all posible info of a request. JS code will be used from now on to determine which inputs will be visible and if its readonly or not. At this point, the JS still needs to be updated.