Prevent receiving account requests with already registered info
Prevent receiving requests for access accounts if the submitted login, email or registration number is the same for some already registered user. But there can be more than one request with the same info, so the admins should know which request accept or archive or ask the requester for the corrected data.