Change deadlines format
Deadlines used to be stored solely as the number of days left, not being possible to specify a different time, only the day number. Change the "chamados" table in the DB to store the deadline as datetime. Also, adapt all necessary code to deal with the new format correctly.

Also:
- It wasn't possible to specify a deadline when accepting a service request, only when a technician assumed them. Now, allow to specify a deadline when accepting a service request, and to update it when assuming one.