Add the request modal UI for managing ticket technicians
The request details modals now include the ability to specify other technicians and what will be their responsibilities with the ticket.
The UI allows seeing editable cards for when technicians still can alter the technicians sharing the ticket, and cards with simpler design for when it's only possible viewing it.